requirements.txt
absl-py==2.4.0
astunparse==1.6.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
filelock==3.24.2
Flask==3.0.3
Flask-Login==0.6.3
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
fsspec==2026.2.0
gast==0.7.0
google-pasta==0.2.0
greenlet==3.3.1
grpcio==1.78.0
gunicorn==25.1.0
h5py==3.15.1
huggingface_hub==0.36.2
idna==3.11
ImageHash==4.3.2
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.13.2
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
mysql-connector-python==9.6.0
mysqlclient==2.2.8
namex==0.1.0
networkx==3.6.1
numpy==1.26.4
opencv-python==4.13.0.92
opencv-python-headless==4.9.0.80
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pillow==10.3.0
protobuf==4.25.8
pycparser==3.0
Pygments==2.19.2
PyMySQL==1.1.2
python-dotenv==1.0.1
PyWavelets==1.9.0
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentencepiece==0.2.0
setuptools==82.0.0
six==1.17.0
SQLAlchemy==2.0.46
sympy==1.14.0
tbb==2021.13.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==3.3.0
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.67.3
transformers==4.40.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.0.3
wheel==0.46.3
wrapt==2.1.1
